Add another backup option for event_decorator.message_id
If the webhook 'msg' attribute does not include an '_id' (this happens on my server occasionally), then we can also read the message id from the actual payload object directly. You will see in the [format documentation](https://mandrill.zendesk.com/hc/en-us/articles/205583307-Message-Event-Webhook-format) that the '_id' field is supposed to be included in both the payload and also nested within the 'msg' field. (And yes, the '_id' refers the email sent, not some "event identifier", as clarified in the documentation.)